refactor: keep cheqd integration out of core config
