+++ src/__tests__/providers/runtime-fetch.test.ts
@@ -2,6 +2,7 @@ import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest';
 import { RuntimeFetchProvider } from '../../providers/runtime-fetch.js';
 import { NodeCryptoProvider } from '../../providers/node-crypto.js';
 import { generateDidKeyFromBase64 } from '../../utils/did-helpers.js';
+import { cheqdResolver } from '../../integrations/cheqd/index.js';
 
 describe('RuntimeFetchProvider', () => {
   let provider: RuntimeFetchProvider;
@@ -37,7 +38,7 @@ describe('RuntimeFetchProvider', () => {
       expect(await provider.resolveDID('did:example:abc')).toBeNull();
     });
 
-    it('returns null for did:cheqd unless a cheqd resolver URL is configured', async () => {
+    it('returns null for did:cheqd unless a cheqd resolver is configured', async () => {
       const fetchSpy = vi.fn();
       vi.stubGlobal('fetch', fetchSpy);
 
@@ -47,10 +48,12 @@ describe('RuntimeFetchProvider', () => {
       expect(fetchSpy).not.toHaveBeenCalled();
     });
 
-    it('resolves did:cheqd through an explicitly configured resolver URL', async () => {
+    it('resolves did:cheqd through an explicitly configured method resolver', async () => {
       const did = 'did:cheqd:testnet:11111111-1111-4111-8111-111111111111';
       const cheqdProvider = new RuntimeFetchProvider({
-        cheqdResolverUrl: 'https://resolver.cheqd.net',
+        didResolvers: {
+          cheqd: cheqdResolver({ resolverUrl: 'https://resolver.cheqd.net' }),
+        },
       });
       const fetchSpy = vi.fn(async (url: string) => {
         expect(url).toBe(`https://resolver.cheqd.net/1.0/identifiers/${did}`);
