+++ src/utils/index.ts
@@ -7,3 +7,6 @@ export {
   ED25519_SPKI_DER_HEADER_LENGTH,
   ED25519_KEY_SIZE,
 } from './ed25519-constants.js';
+
+export { isRecord } from './guards.js';
+export { stripTrailingSlashes } from './url.js';
